BISERVER-11007 - Impossible to use full list of TCP and UDP port numbers while tuning mail server
migrated port type from Short to Integer, added validation to Max and Min values

+++ api/src/org/pentaho/platform/api/email/IEmailConfiguration.java
@@ -38,9 +38,9 @@ public interface IEmailConfiguration {
 
   public void setSmtpHost( final String smtpHost );
 
-  public Short getSmtpPort();
+  public Integer getSmtpPort();
 
-  public void setSmtpPort( final Short smtpPort );
+  public void setSmtpPort( final Integer smtpPort );
 
   public String getSmtpProtocol();
